+++ pkg/auth/api_key.go
@@ -5,7 +5,8 @@ import (
 )
 
 const (
-	GENERATE_API_KEY_ENDPOINT = "system/generate-api-key"
+	GenerateAPIKeyEndpoint       = "system/generate-api-key"
+	GetDatasourcesAPIKeyEndpoint = "system/get-datasources-api-key"
 )
 
 type ApiKey struct {